+++ sdk/search/azure-search-documents/azure/search/documents/_internal/_search_index_document_batching_client.py
@@ -34,10 +34,17 @@ class SearchIndexDocumentBatchingClient(SearchIndexDocumentBatchingClientBase, H
     :param credential: A credential to authorize search client requests
     :type credential: ~azure.core.credentials.AzureKeyCredential
     :keyword bool auto_flush: if the auto flush mode is on. Default to True.
-    :keyword int window: how many seconds if there is no changes that triggers auto flush.
-        Default to 60 seconds
-    :keyword hook: hook. If it is set, the client will call corresponding methods when status changes
-    :paramtype hook: IndexingHook
+    :keyword int auto_flush_interval: how many max seconds if between 2 flushes. This only takes effect
+        when auto_flush is on. Default to 60 seconds
+    :keyword int max_retry_count: total number of retries to allow. Default to 10.
+    :keyword callable new_callback: If it is set, the client will call corresponding methods when there
+        is a new IndexAction added.
+    :keyword callable progress_callback: If it is set, the client will call corresponding methods when there
+        is a IndexAction succeeds.
+    :keyword callable error_callback: If it is set, the client will call corresponding methods when there
+        is a IndexAction fails.
+    :keyword callable remove_callback: If it is set, the client will call corresponding methods when there
+        is a IndexAction removed from the queue (succeeds or fails).
     :keyword str api_version: The Search API version to use for requests.
     """
     # pylint: disable=too-many-instance-attributes
@@ -82,6 +89,7 @@ class SearchIndexDocumentBatchingClient(SearchIndexDocumentBatchingClientBase, H
         """
         return self._index_documents_batch.actions
 
+    @distributed_trace
     def close(self):
         # type: () -> None
         """Close the :class:`~azure.search.documents.SearchClient` session."""
@@ -126,6 +134,8 @@ class SearchIndexDocumentBatchingClient(SearchIndexDocumentBatchingClientBase, H
             except Exception:  # pylint: disable=broad-except
                 pass
 
+        self._reset_timer()
+
         try:
             results = self._index_documents_actions(actions=actions, timeout=timeout)
             for result in results:
@@ -159,9 +169,6 @@ class SearchIndexDocumentBatchingClient(SearchIndexDocumentBatchingClientBase, H
         if not self._auto_flush:
             return
 
-        # reset the timer
-        self._reset_timer()
-
         if len(self._index_documents_batch.actions) < self._batch_size:
             return
 
@@ -173,10 +180,11 @@ class SearchIndexDocumentBatchingClient(SearchIndexDocumentBatchingClientBase, H
             self._timer.cancel()
         except AttributeError:
             pass
-        self._timer = threading.Timer(self._window, self._process)
+        self._timer = threading.Timer(self._auto_flush_interval, self._process)
         if self._auto_flush:
             self._timer.start()
 
+    @distributed_trace
     def add_upload_actions(self, documents):
         # type: (List[dict]) -> None
         """Queue upload documents actions.
@@ -185,9 +193,10 @@ class SearchIndexDocumentBatchingClient(SearchIndexDocumentBatchingClientBase, H
         :type documents: List[dict]
         """
         actions = self._index_documents_batch.add_upload_actions(documents)
-        self._new_callback(actions)
+        self._new_action_callback(actions)
         self._process_if_needed()
 
+    @distributed_trace
     def add_delete_actions(self, documents):
         # type: (List[dict]) -> None
         """Queue delete documents actions
@@ -196,9 +205,10 @@ class SearchIndexDocumentBatchingClient(SearchIndexDocumentBatchingClientBase, H
         :type documents: List[dict]
         """
         actions = self._index_documents_batch.add_delete_actions(documents)
-        self._new_callback(actions)
+        self._new_action_callback(actions)
         self._process_if_needed()
 
+    @distributed_trace
     def add_merge_actions(self, documents):
         # type: (List[dict]) -> None
         """Queue merge documents actions
@@ -207,9 +217,10 @@ class SearchIndexDocumentBatchingClient(SearchIndexDocumentBatchingClientBase, H
         :type documents: List[dict]
         """
         actions = self._index_documents_batch.add_merge_actions(documents)
-        self._new_callback(actions)
+        self._new_action_callback(actions)
         self._process_if_needed()
 
+    @distributed_trace
     def add_merge_or_upload_actions(self, documents):
         # type: (List[dict]) -> None
         """Queue merge documents or upload documents actions
@@ -218,7 +229,7 @@ class SearchIndexDocumentBatchingClient(SearchIndexDocumentBatchingClientBase, H
         :type documents: List[dict]
         """
         actions = self._index_documents_batch.add_merge_or_upload_actions(documents)
-        self._new_callback(actions)
+        self._new_action_callback(actions)
         self._process_if_needed()
 
     def _index_documents_actions(self, actions, **kwargs):
@@ -287,7 +298,7 @@ class SearchIndexDocumentBatchingClient(SearchIndexDocumentBatchingClientBase, H
             # first time that fails
             self._retry_counter[key] = 1
             self._index_documents_batch.enqueue_action(action)
-        elif counter < self._RETRY_LIMIT - 1:
+        elif counter < self._max_retry_count - 1:
             # not reach retry limit yet
             self._retry_counter[key] = counter + 1
             self._index_documents_batch.enqueue_action(action)