+++ sdk/search/azure-search-documents/azure/search/documents/_internal/_search_index_document_batching_client_base.py
@@ -4,39 +4,21 @@
 # license information.
 # --------------------------------------------------------------------------
 from typing import List, TYPE_CHECKING
-from typing_extensions import Protocol
 
 from .._api_versions import validate_api_version
 from .._headers_mixin import HeadersMixin
 
 if TYPE_CHECKING:
     # pylint:disable=unused-import,ungrouped-imports
-    from typing import Any, Union
+    from typing import Any
     from azure.core.credentials import AzureKeyCredential
-
-class IndexingHook(Protocol):
-    def new(self, action, **kwargs):
-        # type: (*str, IndexAction, dict) -> None
-        pass
-
-    def progress(self, action, **kwargs):
-        # type: (*str, IndexAction, dict) -> None
-        pass
-
-    def error(self, action, **kwargs):
-        # type: (*str, IndexAction, dict) -> None
-        pass
-
-    def remove(self, action, **kwargs):
-        # type: (*str, IndexAction, dict) -> None
-        pass
 
 class SearchIndexDocumentBatchingClientBase(HeadersMixin):
     """Base of search index document batching client"""
     _ODATA_ACCEPT = "application/json;odata.metadata=none"  # type: str
-    _DEFAULT_WINDOW = 60
+    _DEFAULT_AUTO_FLUSH_INTERVAL = 60
     _DEFAULT_BATCH_SIZE = 1000
-    _RETRY_LIMIT = 10
+    _MAX_RETRY_COUNT = 10
 
     def __init__(self, endpoint, index_name, credential, **kwargs):
         # type: (str, str, AzureKeyCredential, **Any) -> None
@@ -45,14 +27,18 @@ class SearchIndexDocumentBatchingClientBase(HeadersMixin):
         validate_api_version(api_version)
         self._auto_flush = kwargs.pop('auto_flush', True)
         self._batch_size = kwargs.pop('batch_size', self._DEFAULT_BATCH_SIZE)
-        self._window = kwargs.pop('window', self._DEFAULT_WINDOW)
-        if self._window <= 0:
-            self._window = 86400
+        self._auto_flush_interval = kwargs.pop('auto_flush_interval', self._DEFAULT_AUTO_FLUSH_INTERVAL)
+        if self._auto_flush_interval <= 0:
+            self._auto_flush_interval = 86400
+        self._max_retry_count = kwargs.pop('max_retry_count', self._MAX_RETRY_COUNT)
         self._endpoint = endpoint  # type: str
         self._index_name = index_name  # type: str
         self._index_key = None
         self._credential = credential  # type: AzureKeyCredential
-        self._hook = kwargs.pop('hook', None)
+        self._new_callback = kwargs.pop('new_callback', None)
+        self._progress_callback = kwargs.pop('progress_callback', None)
+        self._error_callback = kwargs.pop('error_callback', None)
+        self._remove_callback = kwargs.pop('remove_callback', None)
         self._retry_counter = {}
 
     @property
@@ -62,18 +48,20 @@ class SearchIndexDocumentBatchingClientBase(HeadersMixin):
 
     def _succeed_callback(self, action):
         # type: (IndexAction) -> None
-        if self._hook:
-            self._hook.remove(action)
-            self._hook.progress(action)
+        if self._remove_callback:
+            self._remove_callback(action)
+        if self._progress_callback:
+            self._progress_callback(action)
 
     def _fail_callback(self, action):
         # type: (IndexAction) -> None
-        if self._hook:
-            self._hook.remove(action)
-            self._hook.error(action)
+        if self._remove_callback:
+            self._remove_callback(action)
+        if self._error_callback:
+            self._error_callback(action)
 
-    def _new_callback(self, actions):
+    def _new_action_callback(self, actions):
         # type: (List[IndexAction]) -> None
-        if self._hook:
+        if self._new_callback:
             for action in actions:
-                self._hook.new(action)
+                self._new_callback(action)