+++ sdk/search/azure-search-documents/samples/sample_batch_client.py
@@ -42,7 +42,6 @@ def sample_batching_client():
             service_endpoint,
             index_name,
             AzureKeyCredential(key),
-            window=100,
             batch_size=100) as batch_client:
         # add upload actions
         batch_client.add_upload_actions(documents=[DOCUMENT])