+++ sdk/search/azure-search-documents/tests/async_tests/test_batching_client_async.py
@@ -17,37 +17,34 @@ CREDENTIAL = AzureKeyCredential(key="test_api_key")
 
 class TestSearchBatchingClientAsync(object):
     async def test_search_index_document_batching_client_kwargs(self):
-        client = SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL, window=100)
-
-        assert client.batch_size == 1000
-        assert client._window == 100
-        assert client._auto_flush
-        await client.close()
+        async with SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL, window=100) as client:
+            assert client.batch_size == 1000
+            assert client._max_retry_count == 10
+            assert client._auto_flush_interval == 60
+            assert client._auto_flush
 
 
     async def test_batch_queue(self):
-        client = SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL, auto_flush=False)
-
-        assert client._index_documents_batch
-        await client.add_upload_actions(["upload1"])
-        await client.add_delete_actions(["delete1", "delete2"])
-        await client.add_merge_actions(["merge1", "merge2", "merge3"])
-        await client.add_merge_or_upload_actions(["merge_or_upload1"])
-        assert len(client.actions) == 7
-        actions = await client._index_documents_batch.dequeue_actions()
-        assert len(client.actions) == 0
-        await client._index_documents_batch.enqueue_actions(actions)
-        assert len(client.actions) == 7
+        async with SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL, auto_flush=False) as client:
+            assert client._index_documents_batch
+            await client.add_upload_actions(["upload1"])
+            await client.add_delete_actions(["delete1", "delete2"])
+            await client.add_merge_actions(["merge1", "merge2", "merge3"])
+            await client.add_merge_or_upload_actions(["merge_or_upload1"])
+            assert len(client.actions) == 7
+            actions = await client._index_documents_batch.dequeue_actions()
+            assert len(client.actions) == 0
+            await client._index_documents_batch.enqueue_actions(actions)
+            assert len(client.actions) == 7
 
 
     @mock.patch(
         "azure.search.documents._internal.aio._search_index_document_batching_client_async.SearchIndexDocumentBatchingClient._process_if_needed"
     )
     async def test_process_if_needed(self, mock_process_if_needed):
-        client = SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL, window=1000, auto_flush=False)
-
-        await client.add_upload_actions(["upload1"])
-        await client.add_delete_actions(["delete1", "delete2"])
+        async with SearchIndexDocumentBatchingClient("endpoint", "index name", CREDENTIAL) as client:
+            await client.add_upload_actions(["upload1"])
+            await client.add_delete_actions(["delete1", "delete2"])
         assert mock_process_if_needed.called